Move artifact service test suite to internal
