+++ artifact/gcs/gcs_test.go
@@ -26,7 +26,7 @@ import (
 
 	"cloud.google.com/go/storage"
 	"google.golang.org/adk/artifact"
-	"google.golang.org/adk/artifact/internal/tests"
+	"google.golang.org/adk/internal/artifact/tests"
 	"google.golang.org/api/iterator"
 )
